Add custom hook useAuth
